backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
pydantic[email]
qdrant-client
google-generativeai
asyncpg
greenlet
requests
itsdangerous
python-multipart
